BROKEN VOXELIZER: u8→u16 palette refactor + GLB pipeline (geometry still wrong)
Ships end-to-end content pipeline from James Catania's voxel-raymarching
(MIT, see tools/scene_voxelize/ATTRIBUTION.md): GLB → GPU compute voxelize
→ k-means in Oklab → .vxs → game tree. Replaces the Python trimesh/
nearest-vertex path, which produced wrong colors.

Also widens the palette from u8 (255 colors, empty at 0) to u16 (65_535
colors, empty lives exclusively on Child::Empty). GpuChild stays 8 bytes
via block_type_lo/hi split. GPU palette moves from a fixed uniform
[vec4;256] to a variable-length storage buffer. Content-AABB cull
(bits 16-27 of packed) dropped since those bits now hold block_type_hi;
shader falls back to the full child bbox. TODO: restore via a parallel
per-BFS aabbs storage buffer.

Sparse .vxs load path added: skips the dense-grid expansion and builds
the tree by walking only occupied voxels — Sponza loads in ~1.3s instead
of iterating 388M padded-cube cells.

**Known-broken:** Sponza voxelization still renders with weird radiating
lines even after Y-up fix. Columns now stand upright (not sideways) but
there's still some geometry bug producing stray thin features. Committing
WIP so we don't lose the palette refactor.

+++ src/world/edit.rs
@@ -96,7 +96,7 @@ pub fn place_child(world: &mut WorldState, hit: &HitInfo, new_child: Child) -> b
 /// Place a block adjacent to the hit face. Builds a uniform subtree
 /// that matches the depth of siblings at the placement site, so the
 /// placed block has full recursive structure like the terrain around it.
-pub fn place_block(world: &mut WorldState, hit: &HitInfo, block_type: u8) -> bool {
+pub fn place_block(world: &mut WorldState, hit: &HitInfo, block_type: u16) -> bool {
     // Figure out how deep siblings are at the placement site.
     let sibling_depth = if let Some(&(parent_id, _)) = hit.path.last() {
         if let Some(parent) = world.library.get(parent_id) {
